MA-5839-3 Implement deactivate data call
One bug was reported by customer that after Wifi is enabled (and then
disabled), 3G can't be used. The root cause is that when Wifi is
enabled, 3G should be shutdown (correct behavior) but we didn't
implement data call deactivation function. As a result, the 3G data
call state in framework level (e.g. MobileDataConnectionTracker) is
not correct.
In RIL, stop PPPd firstly and then deactivate PDP context when
deactivating data call request is received.